+++ autogluon_zeroshot/contexts/utils.py
@@ -1,9 +1,41 @@
-from . import intersect_folds_and_datasets
-from ..simulation.tabular_predictions import TabularPicklePredictions, TabularPicklePredictionsOpt
+from typing import Tuple
+
 from ..simulation.dense_utils import force_to_dense
+from ..simulation.simulation_context import ZeroshotSimulatorContext
+from ..simulation.tabular_predictions import TabularModelPredictions, TabularPicklePredictions, TabularPicklePredictionsOpt
 
 
-def load_zeroshot_input(path_pred_proba, path_gt, zsc, lazy_format: bool = False):
+def intersect_folds_and_datasets(zsc: ZeroshotSimulatorContext,
+                                 zeroshot_pred_proba: TabularModelPredictions,
+                                 zeroshot_gt: dict):
+    zpp_datasets = zeroshot_pred_proba.datasets
+    zsc_datasets = zsc.unique_datasets
+    zsc_datasets_set = set(zsc_datasets)
+    valid_datasets = [d for d in zpp_datasets if d in zsc_datasets_set]
+    if set(zpp_datasets) != set(valid_datasets):
+        zeroshot_pred_proba.restrict_datasets(datasets=valid_datasets)
+        zpp_datasets = zeroshot_pred_proba.datasets
+        zs_gt_keys = zeroshot_gt.keys()
+        for d in zs_gt_keys:
+            if d not in zpp_datasets:
+                zeroshot_gt.pop(d)
+
+    zpp_folds = set(zeroshot_pred_proba.folds)
+    if zpp_folds != set(zsc.folds):
+        zeroshot_pred_proba.restrict_folds(folds=zsc.folds)
+        zs_gt_keys = zeroshot_gt.keys()
+        for d in zs_gt_keys:
+            for f in zpp_folds:
+                if f not in zsc.folds:
+                    zeroshot_gt[d].pop(f)
+    datasets_in_zs = list(zeroshot_pred_proba.datasets)
+    zsc.subset_datasets(datasets_in_zs)
+
+
+def load_zeroshot_input(path_pred_proba: str,
+                        path_gt: str,
+                        zsc: ZeroshotSimulatorContext,
+                        lazy_format: bool = False) -> Tuple[TabularModelPredictions, dict, ZeroshotSimulatorContext]:
     print(f'Loading ZS inputs:\n'
           f'\tpred_proba:  {path_pred_proba}\n'
           f'\tgt:          {path_gt}\n'
