+++ scripts/gen_results/gen_result_from_portfolio.py
@@ -1,7 +1,7 @@
 
 from autogluon.common.savers import save_pd
 
-from autogluon_zeroshot.contexts.context_fetcher import get_context
+from autogluon_zeroshot.contexts import get_context
 from autogluon_zeroshot.simulation.sim_output import SimulationOutputGenerator
 
 
@@ -13,9 +13,9 @@ if __name__ == '__main__':
     This script takes as input a list of configs to generate the output without cross-validation.
     """
 
-    context_name = 'D104_F10_C608_FULL'
+    context_name = 'BAG_D244_F10_C608_FULL'
     benchmark_context = get_context(context_name)
-    name = 'EnsembleAllHPO'
+    name = 'BAG_D244_EnsembleAllHPO'
     zsc, configs_full, zeroshot_pred_proba, zeroshot_gt = benchmark_context.load(load_zpp=True)
     zsc.print_info()
