+++ scripts/method_comparison/run_evaluate_config_lazy.py
@@ -1,5 +1,5 @@
 # File to check that we get the same error as run_evaluate_config.py. It can be removed.
-from autogluon_zeroshot.contexts.context_fetcher import get_context
+from autogluon_zeroshot.contexts import get_context
 from autogluon_zeroshot.utils import catchtime
 from scripts.method_comparison.evaluate_ensemble import evaluate_ensemble
