+++ scripts/run_evaluate_config.py
@@ -1,6 +1,6 @@
 
 from autogluon_zeroshot.simulation.ensemble_selection_config_scorer import EnsembleSelectionConfigScorer
-from autogluon_zeroshot.contexts.context_fetcher import get_context
+from autogluon_zeroshot.contexts import get_context
 from autogluon_zeroshot.utils import catchtime
 
 if __name__ == '__main__':
