+++ scripts/simulate/bagged_289/run_download_input.py
@@ -1,4 +1,4 @@
-from autogluon_zeroshot.contexts.context_fetcher import get_context
+from autogluon_zeroshot.contexts import get_context
 
 
 if __name__ == '__main__':
