+++ scripts/simulate/bagged_289/run_simulate_zs_ensemble_bagged.py
@@ -7,7 +7,7 @@ import numpy as np
 from autogluon.common.savers import save_pkl
 
 from autogluon_zeroshot.simulation.ensemble_selection_config_scorer import EnsembleSelectionConfigScorer
-from autogluon_zeroshot.contexts.context_fetcher import get_context
+from autogluon_zeroshot.contexts import get_context
 from autogluon_zeroshot.simulation.sim_runner import run_zs_simulation, run_zs_simulation_debug
 from autogluon_zeroshot.portfolio import PortfolioCV
 from autogluon_zeroshot.utils import catchtime
