+++ scripts/simulate/bagged_289/run_simulate_zs_single_best_bagged.py
@@ -4,7 +4,7 @@ from autogluon.common.savers import save_pkl
 
 from autogluon_zeroshot.portfolio import PortfolioCV
 from autogluon_zeroshot.simulation.single_best_config_scorer import SingleBestConfigScorer
-from autogluon_zeroshot.contexts.context_fetcher import get_context
+from autogluon_zeroshot.contexts import get_context
 from autogluon_zeroshot.simulation.sim_runner import run_zs_simulation
